backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
sqlalchemy==1.4.23
pydantic==1.8.2
langchain==0.0.184
pypdf==3.8.1
python-dotenv==0.19.0